custom_nodes/comfyui-fishSpeech/requirements.txt
descript-audiotools
dac
hydra-core
tiktoken
loralib
soundfile
pytorch-lightning
lightning
descript-audio-codec
natsort
loguru